refactor: Make the form for new market listing an own component.
